Dependency injection for $skiptoken implementation
